Final SignUp with correct order
